Refactor datastore
