+++ crates/maple_core/src/utils.rs
@@ -1,64 +1,12 @@
-use crate::dirs::PROJECT_DIRS;
 use anyhow::Result;
 use chrono::prelude::*;
 use icon::Icon;
-use std::io::{BufRead, BufReader, Lines};
-use std::path::{Path, PathBuf};
+use std::io::{BufRead, Lines};
+use std::path::Path;
 use subprocess::Exec;
 use utility::{println_json, println_json_with_length, read_first_lines};
 
 pub type UtcTime = DateTime<Utc>;
-
-/// Returns a `PathBuf` using given file name under the project data directory.
-pub fn generate_data_file_path(filename: &str) -> std::io::Result<PathBuf> {
-    let data_dir = PROJECT_DIRS.data_dir();
-    std::fs::create_dir_all(data_dir)?;
-
-    let mut file = data_dir.to_path_buf();
-    file.push(filename);
-
-    Ok(file)
-}
-
-/// Returns a `PathBuf` using given file name under the project cache directory.
-pub fn generate_cache_file_path(filename: impl AsRef<Path>) -> std::io::Result<PathBuf> {
-    let cache_dir = PROJECT_DIRS.cache_dir();
-    std::fs::create_dir_all(cache_dir)?;
-
-    let mut file = cache_dir.to_path_buf();
-    file.push(filename);
-
-    Ok(file)
-}
-
-fn read_json_as<P: AsRef<Path>, T: serde::de::DeserializeOwned>(path: P) -> Result<T> {
-    let file = std::fs::File::open(path)?;
-    let reader = BufReader::new(file);
-    let deserializd = serde_json::from_reader(reader)?;
-
-    Ok(deserializd)
-}
-
-pub fn load_json<T: serde::de::DeserializeOwned, P: AsRef<Path>>(path: Option<P>) -> Option<T> {
-    path.and_then(|json_path| {
-        if json_path.as_ref().exists() {
-            read_json_as::<_, T>(json_path).ok()
-        } else {
-            None
-        }
-    })
-}
-
-pub fn write_json<T: serde::Serialize, P: AsRef<Path>>(
-    obj: T,
-    path: Option<P>,
-) -> std::io::Result<()> {
-    if let Some(json_path) = path.as_ref() {
-        utility::create_or_overwrite(json_path, serde_json::to_string(&obj)?.as_bytes())?;
-    }
-
-    Ok(())
-}
 
 #[derive(Debug, Clone)]
 #[allow(unused)]
